python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.11",
    "grpcio>=1.71.0",
    "protobuf>=6.30.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.2",
]
